Add files via upload
p5.play.js was missing in the index, which is why keyDown was not working.
Radius property was not created inside the circle class
Inappropriate coordinates were given in order to draw an ellipse.
Translate and rotate functions were used. inappropriately in the circle class.
isStatic was missing in the wood class
ground was not created.